Refactor blur (#1259)
* Refactor blur

* Refactor blur

* Fix imports

* Fix tests import

* Fix readme

* Fix mypy

* Fix mypy

* Fix mypy
